get test suite passing with certbot 1.16 (and newer versions)
The test suite failed to run with certbot 1.16 due to

    zope.interface.interfaces.ComponentLookupError: (<InterfaceClass certbot.interfaces.IDisplay>, '')

Since certbot 1.18 the error message got a bit friendlier but the root
cause is basically the same:

   ValueError: This function was called too early in Certbot's execution as the display utility hasn't been configured yet.

We need to use certbot's "patch_display_util()" here. This even works with
certbot 1.1 so we do not need to bump the minimum version in setup.py.